fix 64 to 32 bit clang conversion warning in src/core/lib

+++ src/core/lib/gprpp/bitset.h
@@ -157,7 +157,7 @@ class BitSet {
   ToInt() const {
     Int result = 0;
     for (size_t i = 0; i < kTotalBits; i++) {
-      if (is_set(i)) result |= (Int(1) << i);
+      if (is_set(static_cast<int>(i))) result |= (Int(1) << i);
     }
     return result;
   }
